Extract NMEA-AIS sentence parsing into a package
All logging and threading still happens in the main package.
This commits also adds logging of message types, and hanges some details of Sentence.
A half-finished test of the packet splitting & joining function was completed,
but Sentence parsing, Message assembling and duplicate filtering are still untested.

The next step is to move forwarder and listeners into their own packages (and change listeners to a pull model).