Increase max sentence length
One ship is sending type 5 messages as a single 292-byte sentence.
Type 5 (static and voyage related data) is one of the few types that are
actually parsed, so I think it's worthwile to support it.
The maximum length was already extended from the one in the specification.
The total size of the Sentence struct is unchanged due to previous padding.

With longer sentences more memory could be held by old references in unused
SentenceMerger sentence slots, so they're now cleared when a SentenceMerger
SMID slot is cleared.