Test nmeais.ParseSentence()
I discovered a possible crash if the comma after the payload is the last byte:
The padding is'nt covered by the minimum length.
The bug would never have happened with the output from FirstSentenceInBuffer(),
but ParseSentence() is otherwize independent*, and guarding against it is free.
* Although it have'nt been ran through a fuzzer (yet).

This commit also removes unnecessary casts.