Extract forwarder into a package and test it
* Fixes empty packets not being sent
* makes it possible to stop the forwarder

Until now I had only smoke tested it by looking at the output and seing AIS sentences
fly past, but never checking that the output was correct
Now at least Manager() and forwardTo() is tested with mocks.
The UDP server and Conn impl also needs testing, but it's not a required feature.