Fix double carriage returns from nmeais.FirstSentenceInBuffer()
If a sentence is split between \r and \n, it wouldn't see the previous \r
and insert one. This issue was discovered by feeding the HTTP output
from one server into another.

This commit also removes Some leftover debugging cod, and fixes an existing
test case that didn't pass when it was added (by 80eab316).