Remove broken feature
`for range` copies the value, so only a local copy of fullFor was being incremented.
Fixing that is simple, but reveals another issue:
`forwardTo()` won't find out that the channel is closed before it has emptied the channel,
which might take some time if the client is slow.
This is more complicated to fix, and the brokenness don't appear to have caused any issues.
Because of arbitrary scheduling, a channel being full isn't a good measurement in the first place.